Add overview of doc structure and process to contributing guide
This commit expands the sections on documentation in the CONTRIBUTING.md
file to cover in more depth the new documentation structure and workflow
that is currently being adopted. [1][2] This strives to provide useful
information for so that patch authors and reviewers know how
documentation in ignis is structured and how it gets built as part of
the larger qiskit documentation.

[1] https://github.com/Qiskit/qiskit-terra/pull/3325
[2] https://github.com/Qiskit/qiskit/pull/551